fix: charaterService int 타입 변경된 것 수정
- characterType double로 변경되었던 analyzedDistanceKm int로 다시 되돌려 놓음
- characterService, characterSeriveImpl 기존 코드로 변경

+++ src/main/java/com/smooth/driving_analysis_service/driving/entity/DrivingCharacter.java
@@ -40,7 +40,7 @@ public class DrivingCharacter {
     @Column(name = "description", columnDefinition = "TEXT")
     private String description;
 
-    private Double analyzedDistanceKm;
+    private int analyzedDistanceKm;
 
     private Long fromRecordId;
 
@@ -49,7 +49,7 @@ public class DrivingCharacter {
     @CreationTimestamp
     private LocalDateTime createdAt;
 
-    public void update(DrivingCharacterAnalysisResultDto result, Long toRecordId, double analyzedDistanceKm) {
+    public void update(DrivingCharacterAnalysisResultDto result, Long toRecordId, int analyzedDistanceKm) {
         this.characterType = DrivingCharacterType.valueOf(result.getCharacterType());
         this.confidenceScore = result.getConfidenceScore();
         this.characterTrait = result.getCharacterTrait();
